Extract CLI argument parsing into separate module
